aider: Modified the autocomplete function to show all `availableTags` as suggestions when the "@" character is typed.
# Aider chat conversation:

ASSISTANT: To show all `availableTags` when the "@" character is typed, we need to modify the `assets/static/js/main.js` file.

Currently, the autocomplete function is only triggered when the "@" character is typed, but it doesn't show any suggestions. We need to modify the function to show all `availableTags` as suggestions when the "@" character is typed.

Here is the step-by-step explanation of the changes:

1. We need to modify the `keyup` event listener for the `#search-input` element.
2. Instead of triggering the autocomplete function when the "@" character is typed, we need to trigger it every time a key is pressed.
3. Inside the autocomplete function, we need to check if the input value contains the "@" character. If it does, we show all `availableTags` as suggestions. If it doesn't, we don't show any suggestions.

Here is the *edit block* for the changes:

```javascript
assets/static/js/main.js
<<<<<<< HEAD
    $("#search-input").on('keyup', function(e) {
        if (e.key === "@") {
            console.log("autocomplete");
            $(this).autocomplete({
                source: availableTags
            });
        }
    });
=======
    $("#search-input").on('keyup', function() {
        if (this.value.includes("@")) {
            $(this).autocomplete({
                source: availableTags
            });
        } else {
            $(this).autocomplete({
                source: []
            });
        }
    });
>>>>>>> updated
```

This change will make the autocomplete dropdown show all `availableTags` when the "@" character is typed in the search input field.