aider: Modified the autocomplete function to filter the available tags based on the current input value.
# Aider chat conversation:

ASSISTANT: Yes, it's possible to use a callback function as the `source` option in the autocomplete function. This function will be called every time the user types something in the input field. It receives two arguments: a request object, which has a single `term` property with the current input value, and a response function, which should be called with the filtered array of suggestions.

We can modify the `source` option in the autocomplete function to use a callback function that filters the `availableTags` array based on the current input value. The function should check if the input value starts with the "@" character and if it does, it should filter the `availableTags` array to only include tags that start with the input value (excluding the "@" character). If the input value doesn't start with the "@" character, the function should return an empty array.

Here is the *edit block* for the changes:

```javascript
assets/static/js/main.js
<<<<<<< HEAD
    $("#search-input").on('keyup', function() {
        if (this.value.includes("@")) {
            console.log("email");
            $(this).autocomplete({
                source: availableTags
            });
        } else {
            $(this).autocomplete({
                source: []
            });
        }
    });
=======
    $("#search-input").autocomplete({
        source: function(request, response) {
            if (request.term.startsWith("@")) {
                var term = request.term.slice(1).toLowerCase();
                var suggestions = availableTags.filter(tag => tag.toLowerCase().startsWith(term));
                response(suggestions);
            } else {
                response([]);
            }
        }
    });
>>>>>>> updated
```

This change will make the autocomplete dropdown show all `availableTags` that start with the input value (excluding the "@" character) when the "@" character is typed in the search input field.