aider: Change the `removeButton` into a cross red icon.
# Aider chat conversation:

ASSISTANT: Sure, to change the `removeButton` into a cross red icon, we need to modify the `assets/static/js/main.js` file.

We will replace the text content of the `removeButton` with an HTML string that represents a cross icon. We can use a simple "X" character for this, or if you have a specific icon library in use (like Font Awesome), we can use an icon from there. For this example, I'll use a simple "X" character.

Here is the change:

```javascript
assets/static/js/main.js
<<<<<<< HEAD
            removeButton.textContent = 'Remove';
=======
            removeButton.innerHTML = '<span style="color: red;">X</span>';
>>>>>>> updated
```

Please note that this will simply display a red "X" as the button content. If you want a more sophisticated icon, you'll need to include an icon library like Font Awesome and use an icon from there.